refactor: improve type safety for Meta and BlogPostingParams
- Add discriminated union types: ArticleMeta (with date) and HomeMeta (without date)
- Export BlogPostingParams from types.ts for reuse across modules
- Import BlogPostingParams in jsonld.ts instead of defining locally
- Use 'never' type for date in HomeMeta to enforce type checking
